+++ selene/__init__.py
@@ -220,11 +220,4 @@ Or, by using non-waiting versions, if "you are in a rush:)"::
 # """
 # todo: add here some type imports like Element, Collection, etc.
 
-__version__ = '2.0.0b8'
-
-# --- DEPRECATED, and will be removed soon --- #
-
-
-from selene.support.shared.deprecated import OldConfig as _OldConfig
-
-config = _OldConfig()
+__version__ = '2.0.0b9'