+++ tests/acceptance/acceptance_straightforward_test.py
@@ -37,35 +37,45 @@ is_TodoMVC_loaded = (
 
 class TestTodoMVC:
     def test_selene_demo(self):
-        tasks = browser.all("#todo-list>li")
-        active_tasks = tasks.filtered_by(have.css_class("active"))
+        tasks = browser.all('#todo-list>li')
+        active_tasks = tasks.filtered_by(have.css_class('active'))
 
         browser.open(app_url)
         browser.should(have.js_returned(True, is_TodoMVC_loaded))
 
-        for task_text in ["1", "2", "3"]:
-            browser.element("#new-todo").set_value(task_text).press_enter()
-        tasks.should(have.texts("1", "2", "3")).should_each(
-            have.css_class("active")
+        for task_text in ['1', '2', '3']:
+            browser.element('#new-todo').set_value(task_text).press_enter()
+        tasks.should(have.texts('1', '2', '3')).should_each(
+            have.css_class('active')
         )
-        browser.element("#todo-count").should(have.text('3'))
+        browser.element('#todo-count').should(have.text('3'))
 
-        tasks[2].element(".toggle").click()
-        active_tasks.should(have.texts("1", "2"))
+        tasks[2].element('.toggle').click()
+        active_tasks.should(have.texts('1', '2'))
         active_tasks.should(have.size(2))
 
-        tasks.filtered_by(have.css_class("completed")).should(have.texts("3"))
-        tasks.element_by(not_(have.css_class("completed"))).should(
-            have.text("1")
+        tasks.filtered_by(have.css_class('completed')).should(have.texts('3'))
+        tasks.element_by(not_(have.css_class('completed'))).should(
+            have.text('1')
         )
-        tasks.filtered_by(not_(have.css_class("completed"))).should(
-            have.texts("1", "2")
+        tasks.filtered_by(not_(have.css_class('completed'))).should(
+            have.texts('1', '2')
         )
 
-        browser.element(by.link_text("Active")).click()
-        tasks[:2].should(have.texts("1", "2"))
+        tasks.even.should(have.texts('2'))
+
+        tasks.odd.should(have.texts('1', '3'))
+        # or
+        tasks.sliced(step=2).should(have.texts('1', '3'))
+
+        tasks.sliced().should(have.texts('1', '2', '3'))
+        # or
+        tasks[::].should(have.texts('1', '2', '3'))
+
+        browser.element(by.link_text('Active')).click()
+        tasks[:2].should(have.texts('1', '2'))
         tasks[2].should(be.hidden)
 
-        browser.element(by.id("toggle-all")).click()
-        browser.element("//*[@id='clear-completed']").click()
+        browser.element(by.id('toggle-all')).click()
+        browser.element('//*[@id="clear-completed"]').click()
         tasks.should(have.size(0))